+++ gax-httpjson/src/test/java/com/google/api/gax/httpjson/testing/MockHttpService.java
@@ -27,7 +27,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.google.api.gax.httpjson;
+package com.google.api.gax.httpjson.testing;
 
 import com.google.api.client.http.HttpResponse;
 import com.google.api.client.http.LowLevelHttpRequest;
@@ -35,6 +35,8 @@ import com.google.api.client.http.LowLevelHttpResponse;
 import com.google.api.client.testing.http.MockHttpTransport;
 import com.google.api.client.testing.http.MockLowLevelHttpRequest;
 import com.google.api.client.testing.http.MockLowLevelHttpResponse;
+import com.google.api.gax.httpjson.ApiMessage;
+import com.google.api.gax.httpjson.HttpResponseFormatter;
 import com.google.api.pathtemplate.PathTemplate;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableMap;
